fix: count variants (#147)
## Description

This PR adds a test and a fix for an issue we had when counting variants in the `ToggleCount` class. As reported in #138, the old way we counted didn't work.

I've kept the same testing style as applied in the other metrics bucket tests. Testing 100000 counts seems a little excessive, but it should at least prove that it works. Happy to change it if you think it's too much. 

Fixes #138

## Commits

* chore: test add test that fails to register variant counts

* fix: count variants

* Fix: don't use interlocked for updating a long