fix: increase the possible random numbers used for stickiness id (#241)
This update changes how we calculate random percentages in the SDK. Instead of multiplying a random int with 100 (which absolutely could cause overflows), we now use a ranged `Next` call, to generate one of 10k different numbers. This is in line with how we get the random number in other sdks.

I've added a test to ensure that random distribution works to within the accepted deviation.